Move dependency-resolution-specific code to a separate class; Remove Reset method; simplify internal API;
